+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move Hello World to it's own folder.
 - Part 3 Added a Hello Gophers program.
 - We are going to be building a web services application.
+- Part 5 Added models folder and users model.